API References

API References

Basic Interfaces

Primitive Constructs

Endianness Adapters

Enums

Sequence

Repeaters

Conditional

Padded

Validators

@construct Macro

Construct Sizes

Errors

Basic Interfaces

Basic construct:

Constructs.Construct — Type

Construct{T}

Construct is used for serializing and deserializing objects.

Methods

deserialize(cons::Construct{T}, s::IO; contextkw...)::T

serialize(cons::Construct{T}, s::IO, obj::T; contextkw...)

estimatesize(cons::Construct{T}; contextkw...) - optional

Constructs.serialize — Function

serialize(cons::Construct, s::IO, obj; contextkw...)
serialize(T, s::IO, obj; contextkw...)
serialize(s::IO, obj; contextkw...)

Serialize an object into a stream.

Constructs.deserialize — Function

deserialize(cons::Construct, s::IO; contextkw...)
deserialize(T, s::IO; contextkw...)

Deserialize a stream to an object.

Constructs.estimatesize — Function

estimatesize(cons::Construct; contextkw...)
estimatesize(T; contextkw...)

Estimate the size of the type.

Construct wrapper:

Constructs.Wrapper — Type

Wrapper{TSub, T} <: Construct{T}

Base type of wrapper of TSub.

Methods

subcon(wrapper::Wrapper{TSub, T})::Construct{TSub}

Constructs.subcon — Function

subcon(wrapper::Wrapper{TSub, T})::Construct{TSub}

Get sub-construct of wrapper.

Construct adapter:

Constructs.Adapter — Type

Adapter{TSub, T} <: Wrapper{TSub, T}

Abstract adapter type.

Methods

subcon(adapter::Adapter{TSub, T})::Construct{TSub}

encode(adapter::Adapter{TSub, T}, obj::T; contextkw...)

decode(adapter::Adapter{TSub, T}, obj::TSub; contextkw...)

Constructs.SymmetricAdapter — Type

SymmetricAdapter{T} <: Adapter{T, T}

Abstract adapter type. encode both for serializing and deserializing.

Methods

subcon(adapter::SymmetricAdapter{T})::Construct{T}

encode(adapter::SymmetricAdapter{T}, obj::T; contextkw...)

Constructs.encode — Function

encode(adapter::Adapter{TSub, T}, obj::T; contextkw...) where {TSub, T}

Constructs.decode — Function

decode(adapter::Adapter{TSub, T}, obj::TSub; contextkw...) where {TSub, T}

Construct validator:

Constructs.Validator — Type

Validator{T} <: SymmetricAdapter{T}

Abstract validator type. Validates a condition on the encoded/decoded object..

Methods

subcon(validator::Validator{T})::Construct{T}

validate(validator::Validator{T}, obj::T; contextkw...)

Constructs.validate — Function

validate(validator::Validator{T}, obj::T; contextkw...)

Checks whether the given obj is a valid value for the validator.

Should return a Bool or throw a ValidationError.

Constructs.ValidationError — Type

ValidationError(msg)

Error thrown when the validatiion failed.

Primitive Constructs

Constructs.PrimitiveIO — Type

PrimitiveIO{T} <: Construct{T}

Construct based on primitive read/write.

This is the default construct for Bool, Char, UInt8, UInt16, UInt32, UInt64, UInt128, Int8, Int16, Int32, Int64, Int128, Float16, Float32 and Float64.

Constructs.Singleton — Type

Singleton{T} <: Construct{T}

Singleton type empty construct.

This is the default constructor for Nothing and Missing.

Constructs.JuliaSerializer — Type

JuliaSerializer <: Construct{Any}

Standard Julia serialization.

See also: Serialization

Constructs.RaiseError — Type

RaiseError{E<:Exception} <: Construct{Union{}}

Helper to raise an user-defined error.

Endianness Adapters

Constructs.LittleEndian — Type

LittleEndian{T, TSubCon<:Construct{T}} <: Adapter{T, T}

Little endian data adapter for serializing and deserializing.

Constructs.BigEndian — Type

BigEndian{T, TSubCon<:Construct{T}} <: Adapter{T, T}

Big endian data adapter for serializing and deserializing.

Constructs.Float16be — Constant

Float16be = BigEndian(Float16)

Constructs.Float16le — Constant

Float16le = LittleEndian(Float16)

Constructs.Float32be — Constant

Float32be = BigEndian(Float32)

Constructs.Float32le — Constant

Float32le = LittleEndian(Float32)

Constructs.Float64be — Constant

Float64be = BigEndian(Float64)

Constructs.Float64le — Constant

Float64le = LittleEndian(Float64)

Constructs.Int128be — Constant

Int128be = BigEndian(Int128)

Constructs.Int128le — Constant

Int128le = LittleEndian(Int128)

Constructs.Int16be — Constant

Int16be = BigEndian(Int16)

Constructs.Int16le — Constant

Int16le = LittleEndian(Int16)

Constructs.Int32be — Constant

Int32be = BigEndian(Int32)

Constructs.Int32le — Constant

Int32le = LittleEndian(Int32)

Constructs.Int64be — Constant

Int64be = BigEndian(Int64)

Constructs.Int64le — Constant

Int64le = LittleEndian(Int64)

Constructs.UInt128be — Constant

UInt128be = BigEndian(UInt128)

Constructs.UInt128le — Constant

UInt128le = LittleEndian(UInt128)

Constructs.UInt16be — Constant

UInt16be = BigEndian(UInt16)

Constructs.UInt16le — Constant

UInt16le = LittleEndian(UInt16)

Constructs.UInt32be — Constant

UInt32be = BigEndian(UInt32)

Constructs.UInt32le — Constant

UInt32le = LittleEndian(UInt32)

Constructs.UInt64be — Constant

UInt64be = BigEndian(UInt64)

Constructs.UInt64le — Constant

UInt64le = LittleEndian(UInt64)

Enums

Constructs.IntEnum — Type

IntEnum{Ex<:EnumExhaustibility, T, TSubCon<:Construct{T}, E<:Base.Enum} <: Adapter{T, E}

Integer-based enum adapter for serializing and deserializing.

This is the default constructor for Base.Enum{T}.

Constructs.EnumNonExhaustive — Type

EnumNonExhaustive <: EnumExhaustibility

Indicates the enumeration is non-exhaustive.

Constructs.EnumExhaustive — Type

EnumExhaustive <: EnumExhaustibility

Indicates the enumeration is exhaustive.

Sequence

Constructs.Sequence — Type

Sequence{Tuple{Ts...}} <: Construct{Tuple{Ts...}}

A sequence of construct data.

This is the default constructor for Tuple{Ts...}.

Repeaters

Constructs.SizedArray — Type

SizedArray{T, N, TA<:AbstractArray{T,N}, TSubCon<:Construct{T}} <: Repeater{T, TA}

Homogenous array of elements.

Constructs.GreedyVector — Type

GreedyVector{T, TSubCon<:Construct{T}} <: Repeater{T, Vector{T}}

Homogenous array of elements for unknown count of elements by deserializing until end of stream.

Conditional

Constructs.Try — Type

Try{TU} <: Construct{TU}

Attempts to serialize/deserialize each of the subconstructs.

Padded

Constructs.Padded — Type

Padded{T, TSubCon<:Construct{T}} <: Wrapper{T, T}

Represents Padded data.

Validators

Constructs.Const — Type

Const{T, TSubCon<:Construct{T}} <: Validator{T}

Field enforcing a constant.

@construct Macro

Constructs.@construct — Macro

@construct [ConstructName] structdefinition

Generate a Construct subtype with ConstructName for the given struct.

Examples

julia> @construct struct Bitmap
           ::Const(b"BMP")
           width::UInt16le
           height::UInt16le
           pixel::SizedArray(UInt8, this.width, this.height)
       end

julia> deserialize(Bitmap, b"BMP\x03\x00\x02\x00\x01\x02\x03\x04\x05\x06")
Bitmap(0x0003, 0x0002, UInt8[0x01 0x04; 0x02 0x05; 0x03 0x06])

julia> serialize(Bitmap(2, 3, UInt8[1 2 3; 7 8 9]))
13-element Vector{UInt8}:
 0x42
 0x4d
 0x50
 0x02
 0x00
 0x03
 0x00
 0x01
 0x07
 0x02
 0x08
 0x03
 0x09

julia> estimatesize(Bitmap)
UnboundedSize(0x0000000000000007)

Constructs.this — Constant

this

Placeholder to access properties of the current object in @construct context.

Constructs.Container — Type

Container{T}

Intermediate container for a struct object when serializing/deserializing it.

Container{T}()

Create an uninitialized container for T.

Examples

julia> Container{Complex{Int64}}()
Container{Complex{Int64}}:
  re: #undef
  im: #undef

Constructs.UndefProperty — Type

UndefProperty

Placeholder for undefined properties in Container.

Construct Sizes

Constructs.ConstructSize — Type

ConstructSize

Abstract super type of construct size.

Constructs.ExactSize — Type

ExactSize(value)

Exact construct size (upper bound and lower bound are same).

Constructs.RangedSize — Type

RangedSize(lower, upper)

Ranged construct size.

Constructs.UnboundedSize — Type

UnboundedSize(lower)

Unbounded ranged size.

Constructs.UnboundedUpper — Type

UnboundedUpper

Unsigned infinity.

Errors

Missing docstring.

Missing docstring for ConstructError. Check Documenter's build log for details.

Constructs.ExceedMaxIterations — Type

ExceedMaxIterations(msg, [max_iter])

Error thrown when exceed the max iterations.

Constructs.PaddedError — Type

PaddedError(msg)

Error thrown when the encoded string or bytes takes more bytes than padding allows, or the pad value is improper.