Constructs

Documentation for Constructs.

Constructs.Constructs

Constructs.Float16be

Constructs.Float16le

Constructs.Float32be

Constructs.Float32le

Constructs.Float64be

Constructs.Float64le

Constructs.Int128be

Constructs.Int128le

Constructs.Int16be

Constructs.Int16le

Constructs.Int32be

Constructs.Int32le

Constructs.Int64be

Constructs.Int64le

Constructs.UInt128be

Constructs.UInt128le

Constructs.UInt16be

Constructs.UInt16le

Constructs.UInt32be

Constructs.UInt32le

Constructs.UInt64be

Constructs.UInt64le

Constructs.ValidationOK

Constructs.this

Constructs.Adapter

Constructs.BigEndian

Constructs.BigEndian

Constructs.Const

Constructs.Const

Constructs.Construct

Constructs.Construct

Constructs.ConstructError

Constructs.ConstructSize

Constructs.Container

Constructs.Container

Constructs.EnumExhaustive

Constructs.EnumNonExhaustive

Constructs.ExactSize

Constructs.ExceedMaxIterations

Constructs.GreedyVector

Constructs.GreedyVector

Constructs.IntEnum

Constructs.IntEnum

Constructs.IntEnum

Constructs.JuliaSerializer

Constructs.LittleEndian

Constructs.LittleEndian

Constructs.Padded

Constructs.PaddedError

Constructs.PrimitiveIO

Constructs.RaiseError

Constructs.RaiseError

Constructs.RangedSize

Constructs.Repeater

Constructs.Sequence

Constructs.Singleton

Constructs.SizedArray

Constructs.SizedArray

Constructs.SymmetricAdapter

Constructs.Try

Constructs.UnboundedSize

Constructs.UnboundedUpper

Constructs.UndefProperty

Constructs.ValidationError

Constructs.ValidationOk

Constructs.Validator

Constructs.Wrapper

Constructs.decode

Constructs.deserialize

Constructs.deserialize

Constructs.deserialize

Constructs.encode

Constructs.estimatesize

Constructs.serialize

Constructs.serialize

Constructs.serialize

Constructs.serialize

Constructs.subcon

Constructs.validate

Constructs.@construct

Constructs.Constructs — Module

A declarative de-ser for binary data. Inspired by Construct.

Constructs.Float16be — Constant

Float16be = BigEndian(Float16)

Constructs.Float16le — Constant

Float16le = LittleEndian(Float16)

Constructs.Float32be — Constant

Float32be = BigEndian(Float32)

Constructs.Float32le — Constant

Float32le = LittleEndian(Float32)

Constructs.Float64be — Constant

Float64be = BigEndian(Float64)

Constructs.Float64le — Constant

Float64le = LittleEndian(Float64)

Constructs.Int128be — Constant

Int128be = BigEndian(Int128)

Constructs.Int128le — Constant

Int128le = LittleEndian(Int128)

Constructs.Int16be — Constant

Int16be = BigEndian(Int16)

Constructs.Int16le — Constant

Int16le = LittleEndian(Int16)

Constructs.Int32be — Constant

Int32be = BigEndian(Int32)

Constructs.Int32le — Constant

Int32le = LittleEndian(Int32)

Constructs.Int64be — Constant

Int64be = BigEndian(Int64)

Constructs.Int64le — Constant

Int64le = LittleEndian(Int64)

Constructs.UInt128be — Constant

UInt128be = BigEndian(UInt128)

Constructs.UInt128le — Constant

UInt128le = LittleEndian(UInt128)

Constructs.UInt16be — Constant

UInt16be = BigEndian(UInt16)

Constructs.UInt16le — Constant

UInt16le = LittleEndian(UInt16)

Constructs.UInt32be — Constant

UInt32be = BigEndian(UInt32)

Constructs.UInt32le — Constant

UInt32le = LittleEndian(UInt32)

Constructs.UInt64be — Constant

UInt64be = BigEndian(UInt64)

Constructs.UInt64le — Constant

UInt64le = LittleEndian(UInt64)

Constructs.ValidationOK — Constant

ValidationOK

Placeholder if there is no validatiion error.

Constructs.this — Constant

this

Placeholder to access properties of the current object in @construct context.

Constructs.Adapter — Type

Adapter{TSub, T} <: Wrapper{TSub, T}

Abstract adapter type.

Methods

subcon(adapter::Adapter{TSub, T})::Construct{TSub}

encode(adapter::Adapter{TSub, T}, obj::T; contextkw...)

decode(adapter::Adapter{TSub, T}, obj::TSub; contextkw...)

Constructs.BigEndian — Type

BigEndian{T, TSubCon<:Construct{T}} <: Adapter{T, T}

Big endian data adapter for serializing and deserializing.

Constructs.BigEndian — Method

BigEndian(type)

Defines the big endian format type.

Examples

julia> deserialize(BigEndian(UInt16), b"\x12\x34")
0x1234

Constructs.Const — Type

Const{T, TSubCon<:Construct{T}, VT} <: Validator{T}

Field enforcing a constant.

Constructs.Const — Method

Const([base,] value)

Defines a constant value, usually used for file headers.

Arguments

base::Union{Type, Construct}: the underlying type/construct.

value: the expected value.

Examples

julia> serialize(Const(0x01), 0x01)
1-element Vector{UInt8}:
 0x01

julia> deserialize(Const(0x01), b"\x01")
0x01

julia> serialize(Const(0x01), 0x03)
ERROR: ConstructError: ValidationError: 3 mismatch the const value 1.
[...]

julia> deserialize(Const(0x01), b"\x02")
ERROR: ConstructError: ValidationError: 2 mismatch the const value 1.
[...]

Constructs.Construct — Type

Construct{T}

Construct is used for serializing and deserializing objects.

Methods

deserialize(cons::Construct{T}, s::IO; contextkw...)::T

serialize(cons::Construct{T}, s::IO, obj::T; contextkw...)

estimatesize(cons::Construct{T}; contextkw...) - optional

Constructs.Construct — Method

Construct(type)

Get default construct for type.

Constructs.ConstructError — Type

ConstructError <: Exception

Abstract error type for constructs.

Constructs.ConstructSize — Type

ConstructSize

Abstract super type of construct size.

Constructs.Container — Type

Container{T}

Intermediate container for a struct object when serializing/deserializing it.

Container{T}()

Create an uninitialized container for T.

Examples

julia> Container{Complex{Int64}}()
Container{Complex{Int64}}:
  re: #undef
  im: #undef

Constructs.Container — Method

Container(object)

Create a container from object.

Examples

julia> Container(3+4im)
Container{Complex{Int64}}:
  re: Int64 = 3
  im: Int64 = 4

Constructs.EnumExhaustive — Type

EnumExhaustive <: EnumExhaustibility

Indicates the enumeration is exhaustive.

Constructs.EnumNonExhaustive — Type

EnumNonExhaustive <: EnumExhaustibility

Indicates the enumeration is non-exhaustive.

Constructs.ExactSize — Type

ExactSize(value)

Exact construct size (upper bound and lower bound are same).

Constructs.ExceedMaxIterations — Type

ExceedMaxIterations(msg, [max_iter])

Error thrown when exceed the max iterations.

Constructs.GreedyVector — Type

GreedyVector{T, TSubCon<:Construct{T}} <: Repeater{T, 1, Vector{T}}

Homogenous array of elements for unknown count of elements by deserializing until end of stream.

Constructs.GreedyVector — Method

GreedyVector(element)

Defines an unknown-sized vector, which will deserialize elements as much as possible.

Arguments

element::Union{Type, Construct}: the type/construct of elements.

Constructs.IntEnum — Type

IntEnum{Ex<:EnumExhaustibility, T, TSubCon<:Construct{T}, E<:Base.Enum} <: Adapter{T, E}

Integer-based enum adapter for serializing and deserializing.

Constructs.IntEnum — Method

IntEnum{EnumNonExhaustive}([base,] enum)

Defines the non-exhaustive enum.

Arguments

base::Union{Type, Construct}: the base integer type/construct.

enum::Type: the enum type.

Examples

julia> @enum Fruit::UInt8 apple=1 banana=2 orange=3

julia> deserialize(IntEnum{EnumNonExhaustive}(Fruit), b"\x04")
<invalid #4>::Fruit = 0x04

Constructs.IntEnum — Method

IntEnum([base,] enum)

Defines the (exhaustive) enum.

Arguments

base::Union{Type, Construct}: the base integer type/construct.

enum::Type: the enum type.

Examples

julia> @enum Fruit::UInt8 apple=1 banana=2 orange=3

julia> deserialize(IntEnum(Fruit), b"\x02")
banana::Fruit = 0x02

julia> deserialize(IntEnum(Fruit), b"\x04")
ERROR: ArgumentError: invalid value for Enum Fruit: 4
[...]

julia> serialize(IntEnum(UInt16le, Fruit), orange)
2-element Vector{UInt8}:
 0x03
 0x00

Constructs.JuliaSerializer — Type

JuliaSerializer <: Construct{Any}

Standard Julia serialization.

See also: Serialization

Constructs.LittleEndian — Type

LittleEndian{T, TSubCon<:Construct{T}} <: Adapter{T, T}

Little endian data adapter for serializing and deserializing.

Constructs.LittleEndian — Method

LittleEndian(type)

Defines the little endian format type.

Examples

julia> deserialize(LittleEndian(UInt16), b"\x12\x34")
0x3412

Constructs.Padded — Type

Padded{T, TSubCon<:Construct{T}} <: Wrapper{T, T}

Represents Padded data.

Constructs.PaddedError — Type

PaddedError(msg)

Error thrown when the encoded string or bytes takes more bytes than padding allows, or the pad value is improper.

Constructs.PrimitiveIO — Type

PrimitiveIO{T} <: Construct{T}

Construct based on primitive read/write.

This is the default construct for Bool, Char, UInt8, UInt16, UInt32, UInt64, UInt128, Int8, Int16, Int32, Int64, Int128, Float16, Float32 and Float64.

Constructs.RaiseError — Type

RaiseError{E<:Exception} <: Construct{Union{}}

Helper to raise an user-defined error.

Constructs.RaiseError — Method

RaiseError(error::Exception)
RaiseError(message::String)

Raise specific error or ErrorException(message) when serializing or deserializing any data.

Constructs.RangedSize — Type

RangedSize(lower, upper)

Ranged construct size.

Constructs.Repeater — Type

Repeater{T, N, TA<:AbstractArray{T,N}} <: Wrapper{T, TA}

Abstract base type for Array wrapper.

Constructs.Sequence — Type

Sequence{Tuple{Ts...}} <: Construct{Tuple{Ts...}}

A sequence of construct data.

Constructs.Singleton — Type

Singleton{T} <: Construct{T}

Singleton type empty construct.

Constructs.SizedArray — Type

SizedArray{T, N, TA<:AbstractArray{T,N}, TSubCon<:Construct{T}} <: Repeater{T, N, TA}

Homogenous array of elements.

Constructs.SizedArray — Method

SizedArray([arraytype,] element, size...)

Defines an array with specific size and element.

Arguments

arraytype::Type: the target array type, the default is Array{T, N}.

element::Union{Type, Construct}: the type/construct of elements.

size: the size of the array.

Constructs.SymmetricAdapter — Type

SymmetricAdapter{T} <: Adapter{T, T}

Abstract adapter type. encode both for serializing and deserializing.

Methods

subcon(adapter::SymmetricAdapter{T})::Construct{T}

encode(adapter::SymmetricAdapter{T}, obj::T; contextkw...)

Constructs.Try — Type

Try{TU} <: Construct{TU}

Attempts to serialize/deserialize each of the subconstructs.

Constructs.UnboundedSize — Type

UnboundedSize(lower)

Unbounded ranged size.

Constructs.UnboundedUpper — Type

UnboundedUpper

Unsigned infinity.

Constructs.UndefProperty — Type

UndefProperty

Placeholder for undefined properties in Container.

Constructs.ValidationError — Type

ValidationError(msg)

Error thrown when the validatiion failed.

Constructs.ValidationOk — Type

ValidationOk

Placeholder type if there is no validatiion error.

Constructs.Validator — Type

Validator{T} <: SymmetricAdapter{T}

Abstract validator type. Validates a condition on the encoded/decoded object..

Methods

subcon(validator::Validator{T})::Construct{T}

validate(validator::Validator{T}, obj::T; contextkw...)::Union{ValidationOk, ValidationError}

Constructs.Wrapper — Type

Wrapper{TSub, T} <: Construct{T}

Base type of wrapper of TSub.

Methods

subcon(wrapper::Wrapper{TSub, T})::Construct{TSub}

Constructs.decode — Function

decode(adapter::Adapter{TSub, T}, obj::TSub; contextkw...) where {TSub, T}

Constructs.deserialize — Function

deserialize(cons::Construct, s::IO; contextkw...)
deserialize(T, s::IO; contextkw...)

Deserialize a stream to an object.

Constructs.deserialize — Method

deserialize(cons::Construct, filename::AbstractString; contextkw...)
deserialize(T, filename::AbstractString; contextkw...)

Deserialize a file to an object.

Constructs.deserialize — Method

deserialize(cons::Construct, bytes::AbstractVector{UInt8}; contextkw...)
deserialize(T, bytes::AbstractVector{UInt8}; contextkw...)

Deserialize a byte array to an object.

Constructs.encode — Function

encode(adapter::Adapter{TSub, T}, obj::T; contextkw...) where {TSub, T}

Constructs.estimatesize — Method

estimatesize(cons::Construct; contextkw...)
estimatesize(T; contextkw...)

Estimate the size of the type.

Constructs.serialize — Function

serialize(cons::Construct, s::IO, obj; contextkw...)
serialize(T, s::IO, obj; contextkw...)
serialize(s::IO, obj; contextkw...)

Serialize an object into a stream.

Constructs.serialize — Method

serialize(cons::Construct, filename::AbstractString, obj; contextkw...)
serialize(T, filename::AbstractString, obj; contextkw...)
serialize(filename::AbstractString, obj; contextkw...)

Serialize an object to the file.

Constructs.serialize — Method

serialize(cons::Construct, obj; contextkw...)
serialize(T, obj; contextkw...)
serialize(obj; contextkw...)

Serialize an object in memory (a byte array).

Constructs.serialize — Method

serialize(cons::Construct, s::IO, ::UndefProperty; contextkw...)

Serialize an insufficient object into a stream.

Note

This method is usually called for anonymous fields in @construct.

By default, only singleton types support this because they don't need to write anything.

Constructs.subcon — Method

subcon(wrapper::Wrapper{TSub, T})::Construct{TSub}

Get sub-construct of wrapper.

Constructs.validate — Function

validate(validator::Validator{T}, obj::T; contextkw...)::Union{ValidationOk, ValidationError}

Constructs.@construct — Macro

@construct [ConstructName] structdefinition

Generate a Construct subtype with ConstructName for the given struct.

Examples

```jldoctest julia> @construct struct Bitmap ::Const(b"BMP") width::UInt16le height::UInt16le pixel::SizedArray(UInt8, this.width, this.height) end

julia> deserialize(Bitmap, b"BMP\x03\x00\x02\x00\x01\x02\x03\x04\x05\x06") Bitmap(0x0003, 0x0002, UInt8[0x01 0x04; 0x02 0x05; 0x03 0x06])

julia> serialize(Bitmap(2, 3, UInt8[1 2 3; 7 8 9])) 13-element Vector{UInt8}: 0x42 0x4d 0x50 0x02 0x00 0x03 0x00 0x01 0x07 0x02 0x08 0x03 0x09

julia> estimatesize(Bitmap) UnboundedSize(0x0000000000000007) ````