Constructs

Documentation for Constructs.

Constructs.Constructs

Constructs.ValidationOK

Constructs.this

Constructs.Adapter

Constructs.BaseArray

Constructs.BigEndian

Constructs.Const

Constructs.Construct

Constructs.Construct

Constructs.ConstructError

Constructs.Container

Constructs.ExceedMaxIterations

Constructs.GreedyVector

Constructs.IntEnum

Constructs.JuliaSerializer

Constructs.LittleEndian

Constructs.Padded

Constructs.PaddedError

Constructs.PrimitiveIO

Constructs.Sequence

Constructs.Singleton

Constructs.SizedArray

Constructs.SymmetricAdapter

Constructs.ValidationError

Constructs.ValidationOk

Constructs.Validator

Constructs.Wrapper

Constructs.decode

Constructs.deserialize

Constructs.deserialize

Constructs.deserialize

Constructs.deserialize

Constructs.deserialize

Constructs.deserialize

Constructs.encode

Constructs.estimatesize

Constructs.estimatesize

Constructs.serialize

Constructs.serialize

Constructs.serialize

Constructs.serialize

Constructs.serialize

Constructs.serialize

Constructs.subcon

Constructs.validate

Constructs.@construct

Constructs.Constructs — Module

A declarative de-ser for binary data. Inspired by Construct.

Constructs.ValidationOK — Constant

ValidationOK

Placeholder if there is no validatiion error.

Constructs.this — Constant

this

Placeholder to access properties of the current object in @construct context.

Constructs.Adapter — Type

Adapter{TSub, T} <: Wrapper{TSub, T}

Abstract adapter type.

Methods

subcon(adapter::Adapter{TSub, T})::Construct{TSub}

encode(adapter::Adapter{TSub, T}, obj::T; contextkw...)

decode(adapter::Adapter{TSub, T}, obj::TSub; contextkw...)

Constructs.BaseArray — Type

BaseArray{T, N, TA<:AbstractArray{T,N}} <: Wrapper{T, TA}

Abstract base type for Array wrapper.

Constructs.BigEndian — Type

BigEndian{T, TSubCon<:Construct{T}} <: Adapter{T, T}

Big endian data adapter for serializing and deserializing.

Constructs.Const — Type

Const{T, TSubCon<:Construct{T}, VT} <: Validator{T}

Field enforcing a constant.

Constructs.Construct — Type

Construct{T}

Construct is used for serializing and deserializing objects.

Methods

deserialize(cons::Construct{T}, s::IO; contextkw...)::T

serialize(cons::Construct{T}, s::IO, obj::T; contextkw...)

estimatesize(cons::Construct{T}; contextkw...) - optional

Constructs.Construct — Method

Construct(type)

Get default construct for type.

Constructs.ConstructError — Type

ConstructError <: Exception

Abstract error type for constructs.

Constructs.Container — Type

Container{T}

Intermediate container for a struct object when serializing/deserializing it.

Constructs.ExceedMaxIterations — Type

ExceedMaxIterations(msg, [max_iter])

Error thrown when exceed the max iterations.

Constructs.GreedyVector — Type

GreedyVector{T, TSubCon<:Construct{T}} <: BaseArray{T, 1, Vector{T}}

Homogenous array of elements for unknown count of elements by parsing until end of stream.

Constructs.IntEnum — Type

IntEnum{T, TSubCon<:Construct{T}, E<:Base.Enum} <: Adapter{T, E}

Integer-based enum adapter for serializing and deserializing.

Constructs.JuliaSerializer — Type

JuliaSerializer <: Construct{Any}

Standard Julia serialization.

Constructs.LittleEndian — Type

LittleEndian{T, TSubCon<:Construct{T}} <: Adapter{T, T}

Little endian data adapter for serializing and deserializing.

Constructs.Padded — Type

Padded{T, TSubCon<:Construct{T}} <: Wrapper{T, T}

Represents Padded data.

Constructs.PaddedError — Type

PaddedError(msg)

Error thrown when the encoded string or bytes takes more bytes than padding allows, or the pad value is improper.

Constructs.PrimitiveIO — Type

PrimitiveIO{T} <: Construct{T}

Construct based on primitive read/write.

Constructs.Sequence — Type

Sequence{Tuple{Ts...}} <: Construct{Tuple{Ts...}}

A sequence of construct data.

Constructs.Singleton — Type

Singleton{T} <: Construct{T}

Singleton type empty construct.

Constructs.SizedArray — Type

SizedArray{T, N, TA<:AbstractArray{T,N}, TSubCon<:Construct{T}} <: BaseArray{T, N, TA}

Homogenous array of elements.

Constructs.SymmetricAdapter — Type

SymmetricAdapter{T} <: Adapter{T, T}

Abstract adapter type. encode both for serializing and deserializing.

Methods

subcon(adapter::SymmetricAdapter{T})::Construct{T}

encode(adapter::SymmetricAdapter{T}, obj::T; contextkw...)

Constructs.ValidationError — Type

ValidationError(msg)

Error thrown when the validatiion failed.

Constructs.ValidationOk — Type

ValidationOk

Placeholder type if there is no validatiion error.

Constructs.Validator — Type

Validator{T} <: SymmetricAdapter{T}

Abstract validator type. Validates a condition on the encoded/decoded object..

Methods

subcon(validator::Validator{T})::Construct{T}

validate(validator::Validator{T}, obj::T; contextkw...)::Union{ValidationOK, ValidationError}

Constructs.Wrapper — Type

Wrapper{TSub, T} <: Construct{T}

Base type of wrapper of TSub.

Methods

subcon(wrapper::Wrapper{TSub, T})::Construct{TSub}

Constructs.decode — Function

decode(adapter::Adapter{TSub, T}, obj::TSub; contextkw...) where {TSub, T}

Constructs.deserialize — Function

deserialize(cons::Construct, s::IO; contextkw...)

Deserialize a stream to an object.

Constructs.deserialize — Method

deserialize(cons::Construct, filename::AbstractString; contextkw...)

Deserialize a file to an object.

Constructs.deserialize — Method

deserialize(cons::Construct, bytes::AbstractVector{UInt8}; contextkw...)

Deserialize a byte array to an object.

Constructs.deserialize — Method

deserialize(T, filename::AbstractString; contextkw...)

Deserialize a file to an object.

Constructs.deserialize — Method

deserialize(T, bytes::AbstractVector{UInt8}; contextkw...)

Deserialize a byte array to an object.

Constructs.deserialize — Method

deserialize(T, s::IO; contextkw...)

Deserialize a stream to an object.

Constructs.encode — Function

encode(adapter::Adapter{TSub, T}, obj::T; contextkw...) where {TSub, T}

Constructs.estimatesize — Method

estimatesize(cons::Construct; contextkw...)

Estimate the size of the type.

Constructs.estimatesize — Method

estimatesize(T; contextkw...)

Estimate the size of the type.

Constructs.serialize — Function

serialize(cons::Construct, s::IO, obj; contextkw...)

Serialize an object into a stream.

Constructs.serialize — Method

serialize(filename::AbstractString, obj; contextkw...)

Serialize an object to the file.

Constructs.serialize — Method

serialize(obj; contextkw...)

Serialize an object in memory (a byte array).

Constructs.serialize — Method

serialize(obj, s::IO; contextkw...)

Serialize an object into a stream.

Constructs.serialize — Method

serialize(cons::Construct, filename::AbstractString, obj; contextkw...)

Serialize an object to the file.

Constructs.serialize — Method

serialize(cons::Construct, obj; contextkw...)

Serialize an object in memory (a byte array).

Constructs.subcon — Method

subcon(wrapper::Wrapper{TSub, T})::Construct{TSub}

Get sub-construct of wrapper.

Constructs.validate — Function

validate(validator::Validator{T}, obj::T; contextkw...)::Union{ValidationOk, ValidationError}

Constructs.@construct — Macro

@construct [ConstructName] structdefinition

Generate a Construct{T} subtype with ConstructName for the given struct.